clang: silence -Wcast-align warnings with void* casts
This warning is triggered a lot with clang as the diagnostic is emitted in
the compiler frontend, so the alignment checks and assertions are not
visible to the diagnostic. Silence it by casting to void* for casts
where we know that the value is appropriately aligned.

This is needed to build on FreeBSD with -Werror as the system compiler
is Clang and not GCC.